+++ packages/plugin-packages/mashroom-portal-remote-app-registry-k8s/type-definitions/index.d.ts
@@ -29,6 +29,7 @@ export type KubernetesService = {
     lastCheck: number;
     status: KubernetesServiceStatus;
     error: string | null;
+    retries: number;
     foundPortalApps: Array<MashroomPortalApp>;
     invalidPortalApps: Array<KubernetesServiceInvalidPortalApp>;
 }